+++ doc/changelog.md
@@ -6,7 +6,9 @@ This files is automatically maintained using the ['KornSW-VersioningUtil'](https
 
 *(none)*
 
+- Minor change: ReturnCodeLogger maps >= 0 to "Trace" level instead of "Information"
 
+- **New Feature**: ReturnCodeLogger.Textualize()
 
 ## v 3.0.0
 released **2025-08-13**, including: